move CheckResultsOverview to pages
